code/AzureFunctionsApi/requirements.txt
azure-functions
sqlalchemy>=2.0.0
python-tds>=1.12.0
python-dotenv>=1.0.0

